Arrange topics based on number of replies

Arrange topics based on number of replies

arrange_replies(
  topics,
  by_theme = TRUE,
  by_date = c("month_year", "year", "all")
)

Arguments

topics

A tibble of topics by theme from en-net forum produced through a call to get_themes_topics().

by_theme

Logical. Should topics be grouped by theme? Default is TRUE.

by_date

Should topics be grouped by month of the year or just by year? Default is to group by month of the year.

Value

A tibble of topic views by theme and by specified date format arranged in descending order

Author

Ernest Guevarra

Examples

library(magrittr)
ennet_topics %>% arrange_replies()
#> # A tibble: 1,946 × 9
#>    Theme                Topic  Views Author Posted     Link  Replies Month  Year
#>    <chr>                <chr>  <int> <chr>  <date>     <chr>   <int> <fct> <dbl>
#>  1 Adolescent nutrition Age g…  1601 Anony… 2020-02-15 http…       1 Feb    2020
#>  2 Adolescent nutrition The E…   850 Moham… 2021-02-01 http…       0 Feb    2021
#>  3 Adolescent nutrition Webin…  1434 NATAS… 2021-03-18 http…       5 Mar    2021
#>  4 Adolescent nutrition Webin…  1923 Steph… 2021-03-11 http…       3 Mar    2021
#>  5 Adolescent nutrition Webin…   890 Steph… 2021-03-03 http…       2 Mar    2021
#>  6 Adolescent nutrition Webin…   902 NATAS… 2021-03-02 http…       2 Mar    2021
#>  7 Adolescent nutrition Webin…   774 NATAS… 2021-03-18 http…       1 Mar    2021
#>  8 Adolescent nutrition Webin…   925 NATAS… 2021-03-09 http…       1 Mar    2021
#>  9 Adolescent nutrition Webin…  1040 NATAS… 2021-03-04 http…       1 Mar    2021
#> 10 Adolescent nutrition Webin…   825 NATAS… 2021-03-04 http…       1 Mar    2021
#> # … with 1,936 more rows