Arrange topics based on number of views

Arrange topics based on number of views

arrange_views(
  topics,
  by_theme = TRUE,
  by_date = c("month_year", "year", "all")
)

Arguments

topics
A tibble of topics by theme from en-net forum produced through a call to get_themes_topics().

by_theme
Logical. Should topics be grouped by theme? Default is TRUE.

by_date
Should topics be grouped by month of the year or just by year or overall? Default is to group by month of the year.

Value

A tibble of topic views by theme and by specified date format arranged in descending order

Author

Ernest Guevarra

Examples

library(magrittr)
ennet_topics %>% arrange_views()

#> # A tibble: 3,006 x 9
#>    Theme   Topic          Views Author  Posted     Link      Replies Month  Year
#>    <chr>   <chr>          <int> <chr>   <date>     <chr>       <int> <fct> <dbl>
#>  1 Adoles… Age group of …   599 Anonym… 2020-02-15 https://…       1 Feb    2020
#>  2 Adoles… Launch of res…  1959 Jo Lof… 2018-05-23 https://…       0 May    2018
#>  3 Adoles… Reaching out …  2155 Emily … 2018-06-18 https://…       3 Jun    2018
#>  4 Adoles… adolescent ma…  2120 Anonym… 2018-06-21 https://…       1 Jun    2018
#>  5 Adoles… adolescent ma…  1933 Anonym… 2018-06-21 https://…       1 Jun    2018
#>  6 Adoles… anthropometry…  1890 Emily … 2018-06-29 https://…       2 Jun    2018
#>  7 Adoles… Teen/adolesce…  1827 Anne H… 2018-06-19 https://…       1 Jun    2018
#>  8 Adoles… health educat…  1609 Anonym… 2018-06-19 https://…       1 Jun    2018
#>  9 Adoles… Is there any …   899 Aisha … 2020-06-12 https://…       1 Jun    2020
#> 10 Adoles… Assessing nut…  1912 Ursula… 2018-07-25 https://…       3 Jul    2018
#> # … with 2,996 more rows