Count the number of topics by theme and by date

Count the number of topics by theme and by date

count_topics(topics, by_date = c("month_year", "year"))

Arguments

topics
A tibble of topics by theme from en-net forum produced through a call to get_themes_topics.

by_date
Should topics be counted by month of the year or just by year? Default is by month of the year.

Value

A tibble of topic counts by theme and by specified date format

Examples

library(magrittr)
x <- ennet_topics
x %>% count_topics(by_date = "month_year")

#> # A tibble: 843 x 4
#>    Theme                                Month  Year     n
#>    <chr>                                <fct> <dbl> <int>
#>  1 Adolescent nutrition                 Feb    2020     1
#>  2 Adolescent nutrition                 May    2018     1
#>  3 Adolescent nutrition                 Jun    2018     6
#>  4 Adolescent nutrition                 Jun    2020     1
#>  5 Adolescent nutrition                 Jul    2018     2
#>  6 Adolescent nutrition                 Aug    2018     1
#>  7 Adolescent nutrition                 Sep    2018     1
#>  8 Announcements & Nutritionists needed Jan    2010     1
#>  9 Announcements & Nutritionists needed Jan    2011     4
#> 10 Announcements & Nutritionists needed Jan    2012     3
#> # … with 833 more rows