Count number of questions/topics posted on en-net by time

Count number of questions/topics posted on en-net by time

count_topics_day(topics = get_themes_topics(), .sort = FALSE)

count_topics_week(topics = get_themes_topics(), .sort = FALSE)

count_topics_month(topics = get_themes_topics(), .sort = FALSE)

count_topics_year(topics = get_themes_topics(), .sort = FALSE)

Arguments

topics
A tibble of topics by theme, by author, and by posting date from en-net forum produced through a call to get_themes_topics()

.sort
Logical. Should output be sorted by count frequencies? Default is FALSE

Value

A tibble of topic counts by specified time grouping

Author

Ernest Guevarra

Examples

## Get counts of topics by day
count_topics_day(topics = ennet_topics)

#> # A tibble: 1,806 x 2
#>    day            n
#>  * <date>     <int>
#>  1 2009-02-14     1
#>  2 2009-02-18     1
#>  3 2009-02-19     2
#>  4 2009-02-20     1
#>  5 2009-02-23     1
#>  6 2009-02-24     2
#>  7 2009-02-25     3
#>  8 2009-03-02     1
#>  9 2009-03-03     1
#> 10 2009-03-07     1
#> # … with 1,796 more rows

## Get counts of topics by week
count_topics_week(topics = ennet_topics)

#> # A tibble: 585 x 3
#>    week       week_name        n
#>    <date>     <chr>        <int>
#>  1 2009-02-08 Week 06 2009     1
#>  2 2009-02-15 Week 07 2009     4
#>  3 2009-02-22 Week 08 2009     6
#>  4 2009-03-01 Week 09 2009     3
#>  5 2009-03-08 Week 10 2009     4
#>  6 2009-03-15 Week 11 2009     4
#>  7 2009-03-22 Week 12 2009     2
#>  8 2009-04-12 Week 15 2009     3
#>  9 2009-04-19 Week 16 2009     1
#> 10 2009-04-26 Week 17 2009     4
#> # … with 575 more rows

## Get counts of topics by month
count_topics_month(topics = ennet_topics)

#> # A tibble: 144 x 3
#>    month      month_name     n
#>    <date>     <chr>      <int>
#>  1 2009-02-01 Feb 2009      11
#>  2 2009-03-01 Mar 2009      13
#>  3 2009-04-01 Apr 2009       8
#>  4 2009-05-01 May 2009       8
#>  5 2009-06-01 Jun 2009       8
#>  6 2009-07-01 Jul 2009      10
#>  7 2009-08-01 Aug 2009      10
#>  8 2009-09-01 Sep 2009       8
#>  9 2009-10-01 Oct 2009       2
#> 10 2009-11-01 Nov 2009       4
#> # … with 134 more rows

## Get counts of topics by year
count_topics_year(topics = ennet_topics)

#> # A tibble: 13 x 3
#>    year       year_name     n
#>    <date>     <chr>     <int>
#>  1 2009-01-01 2009         88
#>  2 2010-01-01 2010        103
#>  3 2011-01-01 2011        191
#>  4 2012-01-01 2012        246
#>  5 2013-01-01 2013        228
#>  6 2014-01-01 2014        303
#>  7 2015-01-01 2015        391
#>  8 2016-01-01 2016        362
#>  9 2017-01-01 2017        303
#> 10 2018-01-01 2018        263
#> 11 2019-01-01 2019        254
#> 12 2020-01-01 2020        301
#> 13 2021-01-01 2021         12