Create a daily topics dataset for the ennet_db

Create a daily topics dataset for the ennet_db

create_db_topics_daily(
  gh = "katilingban/ennet_db",
  .date = Sys.Date() - 1,
  fn = NULL
)

create_db_topics_monthly(gh = "katilingban/ennet_db", .date = Sys.Date() - 1)

Arguments

gh
A character value of the GitHub user and repository name combination identifying the GitHub location for ennet_db. Default is katilingban/ennet_db.

.date
A character value or vector of date/dates for which to create a topics dataset for the ennet_db

fn
A character value or vector of filenames for hourly topics dataset found in ennet_db

Author

Ernest Guevarra

Examples

##
fn <- c("ennet_topics_2021-01-17_00:54:48.csv",
        "ennet_topics_2021-01-17_02:48:38.csv",
        "ennet_topics_2021-01-17_03:57:46.csv")

create_db_topics_daily(.date = Sys.Date(), fn = fn)

#> # A tibble: 3,045 x 11
#>    Theme Topic Author Posted Link  Views_20210117_… Replies_2021011…
#>    <chr> <chr> <chr>  <chr>  <chr>            <int>            <int>
#>  1 Anno… "Vac… Anony… 2021-… http…              250               NA
#>  2 Anno… "REA… Saeed… 2021-… http…              139               NA
#>  3 Anno… "New… Katie… 2021-… http…              129               NA
#>  4 Anno… "Eve… Sylvi… 2021-… http…              168               NA
#>  5 Anno… "TAS… Rita … 2021-… http…              142               NA
#>  6 Anno… "ENN… Offic… 2020-… http…              168               NA
#>  7 Anno… "Cal… Tamsi… 2020-… http…              219               NA
#>  8 Anno… "ENN… Offic… 2020-… http…              209               NA
#>  9 Anno… "Vac… Valen… 2020-… http…              355               NA
#> 10 Anno… "Ag-… Sylvi… 2020-… http…              225               NA
#> # … with 3,035 more rows, and 4 more variables: Views_20210117_024838 <int>,
#> #   Replies_20210117_024838 <int>, Views_20210117_035746 <int>,
#> #   Replies_20210117_035746 <int>

create_db_topics_monthly()

#> # A tibble: 3,063 x 727
#>    Theme Topic Author Posted Link  Views_20210101_… Replies_2021010…
#>    <chr> <chr> <chr>  <chr>  <chr>            <int>            <int>
#>  1 Anno… ENN … Offic… 2020-… http…               99               NA
#>  2 Anno… Call… Tamsi… 2020-… http…              127               NA
#>  3 Anno… ENN … Offic… 2020-… http…              135               NA
#>  4 Anno… Vaca… Valen… 2020-… http…              233               NA
#>  5 Anno… Ag-n… Sylvi… 2020-… http…              156               NA
#>  6 Anno… ENN … Offic… 2020-… http…              196               NA
#>  7 Anno… Invi… Offic… 2020-… http…              535               NA
#>  8 Anno… IYCF… Valer… 2020-… http…              212               NA
#>  9 Anno… REAC… Saeed… 2020-… http…              290               NA
#> 10 Anno… IRC … Casie… 2020-… http…              244               NA
#> # … with 3,053 more rows, and 720 more variables: Views_20210101_021536 <int>,
#> #   Replies_20210101_021536 <int>, Views_20210101_032704 <int>,
#> #   Replies_20210101_032704 <int>, Views_20210101_042609 <int>,
#> #   Replies_20210101_042609 <int>, Views_20210101_052611 <int>,
#> #   Replies_20210101_052611 <int>, Views_20210101_061918 <int>,
#> #   Replies_20210101_061918 <int>, Views_20210101_072647 <int>,
#> #   Replies_20210101_072647 <int>, Views_20210101_082054 <int>,
#> #   Replies_20210101_082054 <int>, Views_20210101_092519 <int>,
#> #   Replies_20210101_092519 <int>, Views_20210101_102418 <int>,
#> #   Replies_20210101_102418 <int>, Views_20210101_112233 <int>,
#> #   Replies_20210101_112233 <int>, Views_20210101_121914 <int>,
#> #   Replies_20210101_121914 <int>, Views_20210101_132937 <int>,
#> #   Replies_20210101_132937 <int>, Views_20210101_141710 <int>,
#> #   Replies_20210101_141710 <int>, Views_20210101_151557 <int>,
#> #   Replies_20210101_151557 <int>, Views_20210101_162805 <int>,
#> #   Replies_20210101_162805 <int>, Views_20210101_175049 <int>,
#> #   Replies_20210101_175049 <int>, Views_20210101_183658 <int>,
#> #   Replies_20210101_183658 <int>, Views_20210101_193231 <int>,
#> #   Replies_20210101_193231 <int>, Views_20210101_201950 <int>,
#> #   Replies_20210101_201950 <int>, Views_20210101_211520 <int>,
#> #   Replies_20210101_211520 <int>, Views_20210101_221831 <int>,
#> #   Replies_20210101_221831 <int>, Views_20210101_232633 <int>,
#> #   Replies_20210101_232633 <int>, Views_20210102_003502 <int>,
#> #   Replies_20210102_003502 <int>, Views_20210102_021122 <int>,
#> #   Replies_20210102_021122 <int>, Views_20210102_032824 <int>,
#> #   Replies_20210102_032824 <int>, Views_20210102_042623 <int>,
#> #   Replies_20210102_042623 <int>, Views_20210102_052758 <int>,
#> #   Replies_20210102_052758 <int>, Views_20210102_061942 <int>,
#> #   Replies_20210102_061942 <int>, Views_20210102_072500 <int>,
#> #   Replies_20210102_072500 <int>, Views_20210102_081947 <int>,
#> #   Replies_20210102_081947 <int>, Views_20210102_092349 <int>,
#> #   Replies_20210102_092349 <int>, Views_20210102_102328 <int>,
#> #   Replies_20210102_102328 <int>, Views_20210102_112137 <int>,
#> #   Replies_20210102_112137 <int>, Views_20210102_121745 <int>,
#> #   Replies_20210102_121745 <int>, Views_20210102_132758 <int>,
#> #   Replies_20210102_132758 <int>, Views_20210102_141536 <int>,
#> #   Replies_20210102_141536 <int>, Views_20210102_151401 <int>,
#> #   Replies_20210102_151401 <int>, Views_20210102_162600 <int>,
#> #   Replies_20210102_162600 <int>, Views_20210102_174925 <int>,
#> #   Replies_20210102_174925 <int>, Views_20210102_183549 <int>,
#> #   Replies_20210102_183549 <int>, Views_20210102_193159 <int>,
#> #   Replies_20210102_193159 <int>, Views_20210102_201833 <int>,
#> #   Replies_20210102_201833 <int>, Views_20210102_211337 <int>,
#> #   Replies_20210102_211337 <int>, Views_20210102_221834 <int>,
#> #   Replies_20210102_221834 <int>, Views_20210102_232500 <int>,
#> #   Replies_20210102_232500 <int>, Views_20210103_003707 <int>,
#> #   Replies_20210103_003707 <int>, Views_20210103_021842 <int>,
#> #   Replies_20210103_021842 <int>, Views_20210103_032951 <int>,
#> #   Replies_20210103_032951 <int>, Views_20210103_042655 <int>,
#> #   Replies_20210103_042655 <int>, Views_20210103_052655 <int>,
#> #   Replies_20210103_052655 <int>, …