Create daily topics datasets for the ennet_db

Create daily topics datasets for the ennet_db

create_db_topics_dailies(hourlies)

Arguments

hourlies
A tibble of topics data usually produced by using the create_db_topics_hourlies() function

Value

A tibble of specified topics dataset created from data in the ennet_db

Author

Ernest Guevarra

Examples

themes <- ennet_themes$themes
x <- ennet_hourlies[ennet_hourlies$Theme == themes[3], ]
create_db_topics_dailies(hourlies = x)

#> # A tibble: 840 x 9
#>    Theme Topic Author Posted Link  Interaction Extraction              n
#>    <chr> <chr> <chr>  <chr>  <chr> <chr>       <dttm>              <dbl>
#>  1 COVI… Is t… Umbre… 2020-… http… Views       2020-12-27 23:22:45   174
#>  2 COVI… Is t… Umbre… 2020-… http… Views       2020-12-28 23:22:46   175
#>  3 COVI… Is t… Umbre… 2020-… http… Views       2020-12-29 23:26:27   177
#>  4 COVI… Is t… Umbre… 2020-… http… Views       2020-12-30 23:24:46   186
#>  5 COVI… Is t… Umbre… 2020-… http… Views       2020-12-31 23:24:35   191
#>  6 COVI… Is t… Umbre… 2020-… http… Replies     2020-12-27 23:22:45     0
#>  7 COVI… Is t… Umbre… 2020-… http… Replies     2020-12-28 23:22:46     0
#>  8 COVI… Is t… Umbre… 2020-… http… Replies     2020-12-29 23:26:27     0
#>  9 COVI… Is t… Umbre… 2020-… http… Replies     2020-12-30 23:24:46     0
#> 10 COVI… Is t… Umbre… 2020-… http… Replies     2020-12-31 23:24:35     0
#> # … with 830 more rows, and 1 more variable: `Extraction Date` <date>