Create daily topics dataset for the ennet_db

Create daily topics dataset for the ennet_db

create_db_topics_daily(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date() - 1,
  fn = NULL
)

Arguments

repo

A character value of the GitHub user and repository name combination identifying the GitHub location for ennet_db. Default is katilingban/ennet_db.

branch

A character value for the branch name from which to retrieve data. Default is main.

.date

A character value or vector of date/dates for which to create a topics dataset for the ennet_db

fn

A character value or vector of filenames for hourly topics dataset found in ennet_db

Value

A tibble of daily topics dataset created from data in the ennet_db

Author

Ernest Guevarra

Examples

##
fn <- c("ennet_topics_2021-01-17_00:54:48.csv")

create_db_topics_daily(.date = "2021-01-17", fn = fn)
#> # A tibble: 3,045 × 7
#>    Theme             Topic Author Posted Link  Views_20210117_… Replies_2021011…
#>    <chr>             <chr> <chr>  <chr>  <chr>            <int>            <int>
#>  1 Announcements & … "Vac… Anony… 2021-… http…              250               NA
#>  2 Announcements & … "REA… Saeed… 2021-… http…              139               NA
#>  3 Announcements & … "New… Katie… 2021-… http…              129               NA
#>  4 Announcements & … "Eve… Sylvi… 2021-… http…              168               NA
#>  5 Announcements & … "TAS… Rita … 2021-… http…              142               NA
#>  6 Announcements & … "ENN… Offic… 2020-… http…              168               NA
#>  7 Announcements & … "Cal… Tamsi… 2020-… http…              219               NA
#>  8 Announcements & … "ENN… Offic… 2020-… http…              209               NA
#>  9 Announcements & … "Vac… Valen… 2020-… http…              355               NA
#> 10 Announcements & … "Ag-… Sylvi… 2020-… http…              225               NA
#> # … with 3,035 more rows