Create monthly topics dataset for the ennet_db

Create monthly topics dataset for the ennet_db

create_db_topics_monthly(
  repo = "katilingban/ennet_db",
  branch = "main",
  .date = Sys.Date() - 1
)

Arguments

repo
A character value of the GitHub user and repository name combination identifying the GitHub location for ennet_db. Default is katilingban/ennet_db.

branch
A character value for the branch name from which to retrieve data. Default is main.

.date
A character value or vector of date/dates for which to create a topics dataset for the ennet_db

Value

A tibble of monthly topics dataset created from data in the ennet_db

Author

Ernest Guevarra

Examples

create_db_topics_monthly(.date = "2021-01-03")

#> # A tibble: 3,033 x 143
#>    Theme   Topic       Author   Posted Link    Views_20210101_… Replies_2021010…
#>    <chr>   <chr>       <chr>    <chr>  <chr>              <int>            <int>
#>  1 Adoles… adolescent… Anonymo… 2018-… https:…             2008                1
#>  2 Adoles… adolescent… Anonymo… 2018-… https:…             2197                1
#>  3 Adoles… adolescent… Anonymo… 2018-… https:…             1823                1
#>  4 Adoles… Adolescent… Tamsin … 2018-… https:…             1879                3
#>  5 Adoles… Age group … Anonymo… 2020-… https:…              678                1
#>  6 Adoles… anthropome… Emily M… 2018-… https:…             1971                2
#>  7 Adoles… Assessing … Ursula … 2018-… https:…             1998                3
#>  8 Adoles… health edu… Anonymo… 2018-… https:…             1682                1
#>  9 Adoles… Is there a… Aisha F… 2020-… https:…              982                1
#> 10 Adoles… Launch of … Jo Loft… 2018-… https:…             2046                0
#> # … with 3,023 more rows, and 136 more variables: Views_20210101_021536 <int>,
#> #   Replies_20210101_021536 <int>, Views_20210101_032704 <int>,
#> #   Replies_20210101_032704 <int>, Views_20210101_042609 <int>,
#> #   Replies_20210101_042609 <int>, Views_20210101_052611 <int>,
#> #   Replies_20210101_052611 <int>, Views_20210101_061918 <int>,
#> #   Replies_20210101_061918 <int>, Views_20210101_072647 <int>,
#> #   Replies_20210101_072647 <int>, Views_20210101_082054 <int>,
#> #   Replies_20210101_082054 <int>, Views_20210101_092519 <int>,
#> #   Replies_20210101_092519 <int>, Views_20210101_102418 <int>,
#> #   Replies_20210101_102418 <int>, Views_20210101_112233 <int>,
#> #   Replies_20210101_112233 <int>, Views_20210101_121914 <int>,
#> #   Replies_20210101_121914 <int>, Views_20210101_132937 <int>,
#> #   Replies_20210101_132937 <int>, Views_20210101_141710 <int>,
#> #   Replies_20210101_141710 <int>, Views_20210101_151557 <int>,
#> #   Replies_20210101_151557 <int>, Views_20210101_162805 <int>,
#> #   Replies_20210101_162805 <int>, Views_20210101_175049 <int>,
#> #   Replies_20210101_175049 <int>, Views_20210101_183658 <int>,
#> #   Replies_20210101_183658 <int>, Views_20210101_193231 <int>,
#> #   Replies_20210101_193231 <int>, Views_20210101_201950 <int>,
#> #   Replies_20210101_201950 <int>, Views_20210101_211520 <int>,
#> #   Replies_20210101_211520 <int>, Views_20210101_221831 <int>,
#> #   Replies_20210101_221831 <int>, Views_20210101_232633 <int>,
#> #   Replies_20210101_232633 <int>, Views_20210102_003502 <int>,
#> #   Replies_20210102_003502 <int>, Views_20210102_021122 <int>,
#> #   Replies_20210102_021122 <int>, Views_20210102_032824 <int>,
#> #   Replies_20210102_032824 <int>, Views_20210102_042623 <int>,
#> #   Replies_20210102_042623 <int>, Views_20210102_052758 <int>,
#> #   Replies_20210102_052758 <int>, Views_20210102_061942 <int>,
#> #   Replies_20210102_061942 <int>, Views_20210102_072500 <int>,
#> #   Replies_20210102_072500 <int>, Views_20210102_081947 <int>,
#> #   Replies_20210102_081947 <int>, Views_20210102_092349 <int>,
#> #   Replies_20210102_092349 <int>, Views_20210102_102328 <int>,
#> #   Replies_20210102_102328 <int>, Views_20210102_112137 <int>,
#> #   Replies_20210102_112137 <int>, Views_20210102_121745 <int>,
#> #   Replies_20210102_121745 <int>, Views_20210102_132758 <int>,
#> #   Replies_20210102_132758 <int>, Views_20210102_141536 <int>,
#> #   Replies_20210102_141536 <int>, Views_20210102_151401 <int>,
#> #   Replies_20210102_151401 <int>, Views_20210102_162600 <int>,
#> #   Replies_20210102_162600 <int>, Views_20210102_174925 <int>,
#> #   Replies_20210102_174925 <int>, Views_20210102_183549 <int>,
#> #   Replies_20210102_183549 <int>, Views_20210102_193159 <int>,
#> #   Replies_20210102_193159 <int>, Views_20210102_201833 <int>,
#> #   Replies_20210102_201833 <int>, Views_20210102_211337 <int>,
#> #   Replies_20210102_211337 <int>, Views_20210102_221834 <int>,
#> #   Replies_20210102_221834 <int>, Views_20210102_232500 <int>,
#> #   Replies_20210102_232500 <int>, Views_20210103_003707 <int>,
#> #   Replies_20210103_003707 <int>, Views_20210103_021842 <int>,
#> #   Replies_20210103_021842 <int>, Views_20210103_032951 <int>,
#> #   Replies_20210103_032951 <int>, Views_20210103_042655 <int>,
#> #   Replies_20210103_042655 <int>, Views_20210103_052655 <int>,
#> #   Replies_20210103_052655 <int>, …