Topics from en-net forum retrieved on 17 January 2021

Topics from en-net forum retrieved on 17 January 2021

ennet_topics

Format

A tibble with 3045 rows and 7 columns:

Variable	Description

Theme	Thematic areas in the en-net forum

Topic	Short description of the topic/question being discussed/raised

Views	Number of views of the topic/question being discussed/raised

Author	Name of person who raised the topic/question

Posted	Date topic/question was posted on en-net forum

Link	URL of the topic/question being discussed/raised

Replies	Number of replies to topic/question being discussed/raised

Source

https://www.en-net.org

Details

Please note that this dataset is made available in the package primarily as a guide for the user and as testing data for the code. Users are advised not to use this dataset for actual analysis or reporting and instead make a call to get_theme_topics() or to get_themes_topics(). This is because some topics and some author names have been converted to NA as they contain non-ASCII characters which are not allowed as a text encoding format for packaged data.

Examples

ennet_topics

#> # A tibble: 3,045 x 7
#>    Theme       Topic               Views Author   Posted     Link        Replies
#>    <chr>       <chr>               <int> <chr>    <date>     <chr>         <int>
#>  1 Announceme… "Vacancy - remote …   284 Anonymo… 2021-01-13 https://ww…      NA
#>  2 Announceme… "REACH Initiative …   141 Saeed R… 2021-01-09 https://ww…      NA
#>  3 Announceme… "New Funding Oppor…   132 Katie B… 2021-01-06 https://ww…      NA
#>  4 Announceme… "Event: How and wh…   170 Sylvia … 2021-01-05 https://ww…      NA
#>  5 Announceme… "TASC is looking f…   144 Rita Ab… 2021-01-05 https://ww…      NA
#>  6 Announceme… "ENN is recruiting…   171 Office … 2020-12-23 https://ww…      NA
#>  7 Announceme… "Call for abstract…   220 Tamsin … 2020-12-18 https://ww…      NA
#>  8 Announceme… "ENN is searching …   211 Office … 2020-12-18 https://ww…      NA
#>  9 Announceme… "Vacancy - Nutriti…   359 Valenti… 2020-12-11 https://ww…      NA
#> 10 Announceme… "Ag-nutrition Fund…   227 Sylvia … 2020-12-11 https://ww…      NA
#> # … with 3,035 more rows