Retrieve data from ennet_db GitHub repository

Retrieve data from ennet_db GitHub repository

get_db_discussions(repo = "katilingban/ennet_db", branch = "main")

get_db_topics(
  repo = "katilingban/ennet_db",
  branch = "main",
  id = c("daily", "weekly", "monthly", "yearly")
)

Arguments

repo

A character value for the GitHub user and repository name combination identifying the GitHub location for ennet_db. Default is katilingban/ennet_db.

branch

A character value for the branch name from which to retrieve data. Default is main.

id

A character value for data identifier. Possible choices are daily, weekly, monthly, or yearly.

Value

A tibble of the specified dataset

Author

Ernest Guevarra

Examples

# \donttest{
  ## Retrieve discussions dataset
  get_db_discussions()
#> # A tibble: 9,644 × 10
#>    theme            topic user  userCode job   role  date_time type   code post 
#>    <chr>            <chr> <chr> <chr>    <chr> <chr> <chr>     <chr> <int> <chr>
#>  1 Announcements &… "Reg… Sylv… user404… "Com… Norm… 2022-04-… ques…  4555 "Reg…
#>  2 Announcements &… "Glo… Tams… user24   "en-… Foru… 2022-04-… ques…  4554 "Ple…
#>  3 Announcements &… "Sur… Adam… user409… "Sen… Norm… 2022-04-… ques…  4551 "Dea…
#>  4 Announcements &… "Tac… Coll… user162  ""    Norm… 2022-04-… ques…  4549 "Dea…
#>  5 Announcements &… "Int… Mija… user542  "Joh… Norm… 2022-04-… ques…  4547 "Int…
#>  6 Announcements &… "CON… Jala… user407… "Act… Norm… 2022-04-… ques…  4537 "Clo…
#>  7 Announcements &… "ENN… Offi… user3673 "Adm… Norm… 2022-04-… ques…  4536 "At …
#>  8 Announcements &… "The… Casi… user2155 "IRC" Norm… 2022-03-… ques…  4532 "Job…
#>  9 Announcements &… "The… Casi… user2155 "IRC" Norm… 2022-03-… ques…  4531 "Job…
#> 10 Announcements &… "CON… Jala… user407… "Act… Norm… 2022-03-… ques…  4526 "Clo…
#> # … with 9,634 more rows
# }

## Retrieve en-net topics yearly interactions dataset
get_db_topics(id = "yearly")
#> # A tibble: 53 × 7
#>    Theme        `Extraction Ye…` Replies  Views `New Replies` `New Views` nPosts
#>    <chr>        <date>             <int>  <int>         <int>       <int>  <int>
#>  1 Adolescent … 2020-01-01            19 2.32e4             0           0     13
#>  2 Adolescent … 2021-01-01            38 4.61e4            19       22856     30
#>  3 Adolescent … 2022-01-01            50 1.41e5            12       94833     32
#>  4 Announcemen… 2020-01-01             0 4.72e6             0           0   1241
#>  5 Announcemen… 2021-01-01             0 5.73e6             0     1005126   1340
#>  6 Announcemen… 2022-01-01             0 1.04e7             0     4656472   1474
#>  7 Assessment … 2020-01-01          2097 2.40e6             0           0    456
#>  8 Assessment … 2021-01-01          2539 3.65e6           442     1253011    580
#>  9 Assessment … 2022-01-01          2662 5.79e6           123     2143603    621
#> 10 Coverage as… 2020-01-01           328 5.28e5             0           0     93
#> # … with 43 more rows