Get the discussion and other details for a particular topic

Get the discussion and other details for a particular topic

get_topic_discussions(link)

Arguments

link
URL for topic discussion

Value

A tibble containing the topic question

Examples

links <- get_theme_topics(link = get_themes()$links[4])

#> New names:
#> * `` -> ...1

get_topic_discussions(link = links$Link[1])

#> # A tibble: 12 x 10
#>    theme  topic   user    userCode job     role  date_time type  code  post     
#>    <chr>  <chr>   <chr>   <chr>    <chr>   <chr> <chr>     <chr> <chr> <chr>    
#>  1 Cover… Resour… Tamsin… user24   "en-ne… Foru… 6 Dec 20… ques… 574   "Dear al…
#>  2 Cover… Resour… Mark M… user31   "Consu… Freq… 6 Dec 20… answ… 1536  "And her…
#>  3 Cover… Resour… Ernest… user999  ""      Tech… 22 Dec 2… answ… 1603  "Worksho…
#>  4 Cover… Resour… Saul G… user1000 "Direc… Tech… 11 Jan 2… answ… 1664  "And her…
#>  5 Cover… Resour… Mark M… user31   "Consu… Freq… 1 May 20… answ… 2076  "Just a …
#>  6 Cover… Resour… Ernest… user999  ""      Tech… 26 May 2… answ… 2180  "We are …
#>  7 Cover… Resour… Ernest… user999  ""      Tech… 5 Jul 20… answ… 2278  "Dear Co…
#>  8 Cover… Resour… Mark M… user31   "Consu… Freq… 17 Sep 2… answ… 2442  "Updated…
#>  9 Cover… Resour… Mark M… user31   "Consu… Freq… 1 Nov 20… answ… 2615  "And her…
#> 10 Cover… Resour… Mark M… user31   "Consu… Freq… 30 Apr 2… answ… 3288  "The Cov…
#> 11 Cover… Resour… Mark M… user31   "Consu… Freq… 5 Sep 20… answ… 3445  "Here is…
#> 12 Cover… Resour… Alexan… user107… "Head … Norm… 11 Jun 2… answ… 7723  "The Sta…