Get the discussion and details of discussion for a set of topics

Get the discussion and details of discussion for a set of topics

get_topics_discussions(links)

Arguments

links

A tibble of topics containing URL of topic discussion. This is provided using a call to get_theme_topics() or get_themes_topics()

Value

A tibble containing the topic discussions for selected topic/s

Examples

links <- get_theme_topics(link = (ennet_themes$links)[4])
get_topics_discussions(link = links[1:3, ])
#> # A tibble: 5 × 10
#>   theme   topic user  userCode job   role  date_time           type  code  post 
#>   <chr>   <chr> <chr> <chr>    <chr> <chr> <dttm>              <chr> <chr> <chr>
#> 1 Cross-… How … Shah… user406… "Eco… Norm… 2021-03-26 16:58:00 ques… 4264  "Mor…
#> 2 Cross-… How … Gwén… user409… "Nut… Norm… 2022-04-04 08:55:00 answ… 8513  "A n…
#> 3 Cross-… How … Gius… user3069 "Vet… Norm… 2022-04-04 15:57:00 answ… 8515  "Goo…
#> 4 Cross-… Call… Anon… user24   ""    Foru… 2019-05-10 12:02:00 ques… 3642  "Dea…
#> 5 Cross-… How … Kiro… user225… "Pub… Norm… 2019-05-01 09:40:00 ques… 3637  "Bet…