Retrieve data from ennet_db GitHub repository

Retrieve data from ennet_db GitHub repository

get_db_discussions(repo = "katilingban/ennet_db", branch = "main")

get_db_topics(
  repo = "katilingban/ennet_db",
  branch = "main",
  id = c("daily", "weekly", "monthly", "yearly")
)

Arguments

repo
A character value for the GitHub user and repository name combination identifying the GitHub location for ennet_db. Default is katilingban/ennet_db.

branch
A character value for the branch name from which to retrieve data. Default is main.

id
A character value for data identifier. Possible choices are daily, weekly, monthly, or yearly.

Value

A tibble of the specified dataset

Author

Ernest Guevarra

Examples

## Retrieve discussions dataset
get_db_discussions()

#> # A tibble: 8,823 x 10
#>    theme   topic     user   userCode job    role  date_time type   code post    
#>    <chr>   <chr>     <chr>  <chr>    <chr>  <chr> <chr>     <chr> <int> <chr>   
#>  1 Announ… "Vacancy… Anony… user405… "Nutr… Norm… 2021-01-… ques…  4181 "We are…
#>  2 Announ… "REACH I… Saeed… user405… "Nutr… Norm… 2021-01-… ques…  4177 "REACH …
#>  3 Announ… "New Fun… Katie… user404… "Tech… Norm… 2021-01-… ques…  4174 "I am h…
#>  4 Announ… "Event: … Sylvi… user404… "Comm… Norm… 2021-01-… ques…  4172 "Join t…
#>  5 Announ… "TASC is… Rita … user405… "Nutr… Norm… 2021-01-… ques…  4171 "Progre…
#>  6 Announ… "ENN is … Offic… user3673 "Admi… Norm… 2020-12-… ques…  4167 "We are…
#>  7 Announ… "Call fo… Tamsi… user24   "en-n… Foru… 2020-12-… ques…  4164 "The fo…
#>  8 Announ… "ENN is … Offic… user3673 "Admi… Norm… 2020-12-… ques…  4163 "ENN is…
#>  9 Announ… "Vacancy… Valen… user404… "Glob… Norm… 2020-12-… ques…  4156 "Nutrit…
#> 10 Announ… "Ag-nutr… Sylvi… user404… "Comm… Norm… 2020-12-… ques…  4155 "The In…
#> # … with 8,813 more rows

## Retrieve en-net topics daily interactions dataset
get_db_topics(id = "daily")

#> # A tibble: 378 x 7
#>    Theme         `Extraction Dat… Replies Views `New Replies` `New Views` nPosts
#>    <chr>         <date>             <int> <int>         <int>       <int>  <int>
#>  1 Adolescent n… 2020-12-27            19 23059             0           0     13
#>  2 Adolescent n… 2020-12-28            19 23084             0          25     13
#>  3 Adolescent n… 2020-12-29            19 23125             0          41     13
#>  4 Adolescent n… 2020-12-30            19 23177             0          52     13
#>  5 Adolescent n… 2020-12-31            19 23249             0          72     13
#>  6 Adolescent n… 2021-01-01            19 23275             0          26     13
#>  7 Adolescent n… 2021-01-02            19 23301             0          26     13
#>  8 Adolescent n… 2021-01-03            19 23325             0          24     13
#>  9 Adolescent n… 2021-01-04            19 23348             0          23     13
#> 10 Adolescent n… 2021-01-05            19 23385             0          37     13
#> # … with 368 more rows